% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/candidate_pairs.R
\name{candidate_pairs}
\alias{candidate_pairs}
\title{Candidate pairs}
\usage{
candidate_pairs(m)
}
\arguments{
\item{m}{A matrix from \code{\link{pairwise_cf}}.}
}
\value{
A data frame containing all the non-\code{NA} values from \code{m}.
  Columns \code{a} and \code{b} are the IDs from the original corpus as
  passed to the comparison function. Column \code{score} is the score
  returned by the comparison function.
}
\description{
Converts a comparison matrix generated by \code{\link{pairwise_cf}} into a
data frame of candidates for matches.
}
\examples{
dir <- system.file("extdata", package = "textreuse")
corpus <- TextReuseCorpus(dir = dir)
names(corpus) <- filenames(names(corpus))
m <- pairwise_cf(corpus, ratio_of_matches, directional = TRUE)
candidate_pairs(m)
}

