% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/corpus_cf.R
\name{corpus_cf}
\alias{corpus_cf}
\title{Compare documents in a corpus to one another}
\usage{
corpus_cf(x, y, f, ..., id = "file")
}
\arguments{
\item{x}{A list of documents to be compared.}

\item{y}{A list of documents to be compared. Often \code{y} will be the same
as \code{x}.}

\item{f}{The function to apply to \code{x} and \code{y}.}

\item{...}{Additional arguments passed to \code{f}.}

\item{id}{The name of the value in the \code{meta} list of the
  \code{\link{TextReuseTextDocument}} to be used in row and column names of
  the matrix. Set to \code{NULL} to remove row and column names.}
}
\description{
Given two lists containing documents of class
\code{\link{TextReuseTextDocument}}, this function applies a comparison
function to every pairing of documents, and returns a matrix with the
comparison scores.
}
\examples{
ny         <- system.file("extdata/ny1850-match.txt", package = "textreuse")
ca_match   <- system.file("extdata/ca1851-match.txt", package = "textreuse")
ca_nomatch <- system.file("extdata/ca1851-nomatch.txt", package = "textreuse")

ny         <- TextReuseTextDocument(ny, meta = list(id = "ny"))
ca_match   <- TextReuseTextDocument(ca_match, meta = list(id = "ca_match"))
ca_nomatch <- TextReuseTextDocument(ca_nomatch, meta = list(id = "ca_no"))

corpus <- list(ny, ca_match, ca_nomatch)

corpus_cf(corpus, corpus, jaccard_similarity, id = "id")
corpus_cf(corpus, corpus, ratio_of_matches, id = "id")
}
\seealso{
See these document comparison functions, \code{\link{jaccard_similarity}},
  \code{\link{ratio_of_matches}}.
}

