% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenames.R
\name{filenames}
\alias{filenames}
\title{Filenames from paths}
\usage{
filenames(paths, extension = FALSE)
}
\arguments{
\item{paths}{A character vector of paths.}

\item{extension}{Should the file extension be preserved?}
}
\description{
This function takes a character vector of paths and returns just the file
name, by default without the extension. A \code{\link{TextReuseCorpus}} uses
the paths to the files in the corpus as the names of the list. This function
is intended to turn those paths into more manageable identifiers.
}
\examples{
paths <- c("corpus/one.txt", "corpus/two.md", "corpus/three.text")
filenames(paths)
filenames(paths, extension = TRUE)
}
\seealso{
\code{\link{basename}}
}
