% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hash_string}
\alias{hash_string}
\title{Hash a string to an integer}
\usage{
hash_string(x)
}
\arguments{
\item{x}{A character vector to be hashed.}
}
\value{
A vector of integer hashes.
}
\description{
Hash a string to an integer
}
\examples{
s <- c("How", "many", "roads", "must", "a", "man", "walk", "down")
hash_string(s)
}
