% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{is.TextReuseTextDocument}
\alias{is.TextReuseTextDocument}
\title{Check if an object is a \code{\link{TextReuseTextDocument}}.}
\usage{
is.TextReuseTextDocument(x)
}
\arguments{
\item{x}{An R object to check.}
}
\description{
Check if an object is a \code{\link{TextReuseTextDocument}}.
}

