% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jaccard.R
\name{jaccard}
\alias{jaccard}
\alias{jaccard_bag_similarity}
\alias{jaccard_dissimilarity}
\alias{jaccard_similarity}
\title{Calculate Jaccard similarity/dissimilarity coefficients}
\usage{
jaccard_similarity(a, b)

jaccard_dissimilarity(a, b)

jaccard_bag_similarity(a, b)
}
\arguments{
\item{a}{The first set (or bag) to be compared.}

\item{b}{The second set (or bag) to be compared.}
}
\description{
These functions compute the Jaccard measures of similarity or dissimilarity
for two sets. The coefficients will be numbers between \code{0} and \code{1}.
For the similarity coefficient, the higher the number the more similar the
two sets are. When applied to two documents of class
\code{\link{TextReuseTextDocument}}, the n-grams in those documents are
compared. But this function can be passed objects of any class accepted by
the set functions in base R. So it is possible, for instance, to pass this
function two character vectors comprised of word, line, sentence, or
paragraph tokens, or those character vectors hashed as integers.
}
\details{
The Jaccard similarity coeffecient is defined as follows:

\deqn{J(A, B) = \frac{ | A \cup B | }{ | A \cap B | }}{ length(intersect(a,
b)) / length(union(a, b))}

The dissimilarity is simply

\deqn{1 - J(A, B)}

The Jaccard bag similarity treats \code{a} and \code{b} as bags rather than
sets, so that the result is a fraction where the numerator is the sum of each
matching element counted the minimum number of times it appears in each bag,
and the denominator is the sum of the lengths of both bags. The maximum value
for the Jaccard bag similarity is \code{0.5}.
}
\examples{
jaccard_similarity(1:3, 2:4)
jaccard_dissimilarity(1:3, 2:4)

ny         <- system.file("extdata/ny1850-match.txt", package = "textreuse")
ca_match   <- system.file("extdata/ca1851-match.txt", package = "textreuse")
ca_nomatch <- system.file("extdata/ca1851-nomatch.txt", package = "textreuse")

ny         <- TextReuseTextDocument(ny, meta = list(id = "ny"))
ca_match   <- TextReuseTextDocument(ca_match, meta = list(id = "ca_match"))
ca_nomatch <- TextReuseTextDocument(ca_nomatch, meta = list(id = "ca_nomatch"))

# These two should have a higher Jaccard similarity coefficient
jaccard_similarity(ny, ca_match)

# These two should have a lower Jaccard similarity coefficient
jaccard_similarity(ny, ca_nomatch)

a <- c("a", "a", "a", "b")
b <- c("a", "a", "b", "b", "c")
jaccard_similarity(a, b)
jaccard_bag_similarity(a, b)
}
\references{
Jure Leskovec, Anand Rajaraman, Jeff Ullman,
  \href{http://www.mmds.org/#book}{\emph{Mining of Massive Datasets}}
  (Cambridge University Press, 2011).
}

