% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jaccard.R
\name{jaccard_similarity}
\alias{jaccard_similarity}
\title{Calculate Jaccard similarity coefficient}
\usage{
jaccard_similarity(a, b)
}
\arguments{
\item{a}{The first set to be compared.}

\item{b}{The second set to be compared.}
}
\description{
This function computes a measure of similarity for two sets. The higher the
number, the more similar the two sets are. When applied to two documents of
class \code{\link{TextReuseTextDocument}}, the n-grams in those documents are
compared. But this function can be passed sets of any class accepted by the
set functions in base R. So it is possible, for instance, to pass this
function two character vectors comprised of word, line, sentence, or
paragraph tokens, or those character vectors hashed as integers.
}
\details{
The Jaccard similarity coeffecient is defined as follows:

\deqn{J(A, B) = \frac{ | A \cup B | }{ | A \cap B | }}{ length(intersect(a,
b)) / length(union(a, b))}
}
\examples{
jaccard_similarity(1:3, 2:4)

ny         <- system.file("extdata/ny1850-match.txt", package = "textreuse")
ca_match   <- system.file("extdata/ca1851-match.txt", package = "textreuse")
ca_nomatch <- system.file("extdata/ca1851-nomatch.txt", package = "textreuse")

ny         <- TextReuseTextDocument(ny)
ca_match   <- TextReuseTextDocument(ca_match)
ca_nomatch <- TextReuseTextDocument(ca_nomatch)

# These two should have a higher Jaccard similarity coefficient
jaccard_similarity(ny, ca_match)

# These two should have a lower Jaccard similarity coefficient
jaccard_similarity(ny, ca_nomatch)
}

