% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsh.R
\name{lsh_add}
\alias{lsh_add}
\title{Add documents to a LSH cache}
\usage{
lsh_add(buckets, x, bands, progress = interactive())
}
\arguments{
\item{buckets}{An \code{lsh_buckets} object created by \code{\link{lsh}}.}

\item{x}{A \code{\link{TextReuseCorpus}} or
\code{\link{TextReuseTextDocument}} with minhashes.}

\item{bands}{The number of bands to use for locality sensitive hashing. The
number of hashes in the documents in the corpus must be evenly divisible by
the number of bands. See \code{\link{lsh_threshold}} and
\code{\link{lsh_probability}} for guidance in selecting the number of bands
and hashes.}

\item{progress}{Display a progress bar while comparing documents.}
}
\value{
An updated \code{lsh_buckets} object.
}
\description{
This function adds buckets for one or more new documents to an existing
\code{lsh_buckets} object. Use the same \code{bands} value and minhash
function that were used to create the original buckets.
}
\seealso{
\code{\link{lsh}}, \code{\link{lsh_query}},
  \code{\link{lsh_candidates}}
}
