% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/minhash.R
\name{minhash}
\alias{minhash}
\title{Calculate minhash values}
\usage{
minhash(x, random)
}
\arguments{
\item{x}{The vector to be minhashed.}

\item{random}{A vector of integers.}
}
\value{
An integer vector with the same length as \code{random}.
}
\description{
A minhash value is calculated by hashing a string to an integer and selecting
the minimum value. Repeated minhash values are generated by using different
hash functions: these different hash functions are created by using
performing a bitwise \code{XOR} operation (\code{\link{bitwXor}}) with a
vector of random integers. It is vital that you use the same random integers
for every document in your corpus. To generate these, see
\code{\link{random_ints}}.
}
\examples{
set.seed(253)
r <- random_ints(3)
minhash(c("Some", "words", "are", "the", "same"), r)
minhash(c("Some", "words", "are", "different", "though"), r)
}
\seealso{
\code{\link{random_ints}}
}

