% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ngrams.R
\name{ngrams}
\alias{ngrams}
\title{Break a text into n-grams}
\usage{
ngrams(string, n = 3, lowercase = TRUE)
}
\arguments{
\item{string}{A character vector to be split into n-grams.}

\item{n}{The value of \code{n}.}

\item{lowercase}{Should the n-grams be made lowercase.}
}
\value{
A character vector containing the n-grams.
}
\description{
Break a text into n-grams
}
\details{
As a side-effect this function strips punctuation.
}

