% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ocr_quality.R
\name{ocr_quality}
\alias{ocr_quality}
\title{Measure the quality of an text created with OCR}
\usage{
ocr_quality(text, method = c("dictionary"), sample_size = -1L)
}
\arguments{
\item{text}{A character vector. Each text (or chunk of a text) should be a
value within the vector.}

\item{method}{The method used. \describe{ \item{\code{dictionary}}{Checks
whether the words used in the text are found in a list of English words. It
cannot identify OCR errors that result in the wrong English word. (See
\code{\link{words_en}}.) Returns the ratio of words found in the list to
the total number of words in the text.}}}

\item{sample_size}{If this value is positive, then this many words from the
\code{text} will be selected for comparison. This is useful for large
texts.}
}
\value{
A vector of numeric values between \code{0} and \code{1}.
}
\description{
This function checks the quality of an OCR text. It will return a number
between \code{0} and \code{1}. The higher the number, the better the quality
of the OCR. These measures should not be taken in an absolute sense. That is,
a score of 1 does not indicate perfect OCR. They should only be used to
determine the relative quality of OCR within a corpus of texts. See the
description of methods used to calculate OCR quality below. You can pass a
character vector of any length. So, if you split a text into chunks, you can
evaluate the OCR quality of each chunk.
}
\examples{
paragraph <- "Fourr score and sleven years ago our fathers brought
  forth on this continent, a new nation, conceived in Liberty,
  and dedicated to tlhe proposition that all men are created equal."

ocr_quality(paragraph)
}

