% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/minhash.R
\name{random_ints}
\alias{random_ints}
\title{Generate random integers for minhashing}
\usage{
random_ints(n)
}
\arguments{
\item{n}{The number of random integers to generate.}
}
\value{
A vector of integers
}
\description{
It is crucial that you use the same random integers for every document in the
corpus. The random integers generated by this function are intended to be
passed to \code{\link{minhash}}.
}
\examples{
random_ints(3)
}
\seealso{
\code{\link{minhash}}
}

