% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ratio_of_matches.R
\name{ratio_of_matches}
\alias{ratio_of_matches}
\title{Ratio of matches in two vectors or documents}
\usage{
ratio_of_matches(origin, destination)
}
\arguments{
\item{origin}{The vector which is expected to be the origin.}

\item{destination}{The vector which is expected to be the destination}
}
\description{
This function takes two vectors or documents, \code{origin} and
\code{destination}. It finds the ratio between the number of items in
\code{destination} that are also in \code{origin} and the total number of
items in \code{destination}. Note that this similarity measure is
directional: it measures how much \code{destination} borrows from
\code{origin}, but says nothing about how much of \code{origin} borrows from
destination. If this function is passed two documents of class
\code{\link{TextReuseTextDocument}}, it will compute the similarity using the
n-grams in those documents.
}
\examples{
ratio_of_matches(1:3, 2:4)

ny         <- system.file("extdata/ny1850-match.txt", package = "textreuse")
ca_match   <- system.file("extdata/ca1851-match.txt", package = "textreuse")
ca_nomatch <- system.file("extdata/ca1851-nomatch.txt", package = "textreuse")

ny         <- TextReuseTextDocument(ny)
ca_match   <- TextReuseTextDocument(ca_match)
ca_nomatch <- TextReuseTextDocument(ca_nomatch)

# These two should have a higher ratio of matches
ratio_of_matches(ny, ca_match)

# These two should have a very low ratio of matches
ratio_of_matches(ny, ca_nomatch)
}

