% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextReuseTextDocument.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{meta}
\alias{meta<-}
\alias{content}
\alias{content<-}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{NLP}{\code{\link[NLP:content]{content()}}, \code{\link[NLP:content<-]{content<-()}}, \code{\link[NLP:meta]{meta()}}, \code{\link[NLP:meta<-]{meta<-()}}}
}}

