% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/textreuse-package.r
\docType{package}
\name{textreuse}
\alias{textreuse}
\alias{textreuse-package}
\title{textreuse: Detect Text Reuse and Document Similarity}
\description{
textreuse: Detect Text Reuse and Document Similarity
}

