% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tr_document.R
\name{tr_document}
\alias{tr_document}
\title{Create an document object}
\usage{
tr_document(string, file = NULL, n = 5, ...)
}
\arguments{
\item{string}{A character vector of length 1.}

\item{file}{The path to a file, if \code{string} is not provided.}

\item{n}{The value of \code{n} in the n-grams that are created by
\link{ngrams}.}

\item{...}{Arguments are passed on to \link{readLines} if \code{file} is
  provided.}
}
\value{
An object of class \code{tr_document}. This object inherits from
  \code{list}, and contains the following: \code{text} (the document text),
  \code{file} (the path to the file), \code{ngrams} (the n-grams created from
  the document).
}
\description{
This function either reads in a character vector or a file and creates an
object of class \code{tr_document}. This class is used for comparing
documents.
}

