% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordcount.R
\name{wordcount}
\alias{wordcount}
\title{Count words}
\usage{
wordcount(x)
}
\arguments{
\item{x}{The object containing a text.}
}
\value{
An integer vector for the word count.
}
\description{
This function counts words in a text, for example, a character vector, a
\code{\link{TextReuseTextDocument}}, some other object that inherits from
\code{\link[NLP]{TextDocument}}, or a all the documents in a
\code{\link{TextReuseCorpus}}.
}
