% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lsh_candidates.R
\name{lsh_candidates}
\alias{lsh_candidates}
\title{Candidate pairs from LSH comparisons}
\usage{
lsh_candidates(buckets, pairs = TRUE)
}
\arguments{
\item{buckets}{A cache object returned from \code{\link{lsh}}.}

\item{pairs}{If \code{TRUE}, pairs of candidates will be returned. Otherwise,
clusters of candidates will be returned.}
}
\value{
A list containing character vectors with the IDs of potential matches.
}
\description{
Given a cache object from an LSH comparison, this function returns the
potential candidates.
}
\examples{
dir <- system.file("extdata", package = "textreuse")
set.seed(234)
minhash2 <- minhash_generator(200)
corpus <- TextReuseCorpus(dir = dir,
                          tokenizer = tokenize_ngrams, n = 5,
                          hash_func = minhash2)
names(corpus) <- filenames(names(corpus))
buckets <- lsh(corpus, bands = 50)
lsh_candidates(buckets)
}

