% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lsh_candidates.R
\name{lsh_candidates}
\alias{lsh_candidates}
\title{Candidate pairs from LSH comparisons}
\usage{
lsh_candidates(buckets)
}
\arguments{
\item{buckets}{A cache object returned from \code{\link{lsh}}.}
}
\value{
A list containing character vectors with the IDs of potential matches.
}
\description{
Given a cache object from an LSH comparison, this function returns the
potential candidates.
}
\examples{
dir <- system.file("extdata/legal", package = "textreuse")
minhash <- minhash_generator(200, seed = 234)
corpus <- TextReuseCorpus(dir = dir,
                          tokenizer = tokenize_ngrams, n = 5,
                          hash_func = minhash)
buckets <- lsh(corpus, bands = 50)
lsh_candidates(buckets)
}

