% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rehash.R
\name{rehash}
\alias{rehash}
\title{Recompute the hashes for a document or corpus}
\usage{
rehash(x, hash_func)
}
\arguments{
\item{x}{A \code{\link{TextReuseTextDocument}} or
\code{\link{TextReuseCorpus}}.}

\item{hash_func}{A function to hash the tokens. See
\code{\link{hash_string}}, \code{\link{minhash_generator}}.}
}
\value{
The modified \code{\link{TextReuseTextDocument}} or
  \code{\link{TextReuseCorpus}}.
}
\description{
Given a \code{\link{TextReuseTextDocument}} or a
\code{\link{TextReuseCorpus}}, this function recomputes the hashes with the
function specified. This implies that you have retained the tokenens with the
\code{keep_tokens = TRUE} parameter. This function is intended to be used
which changing from keeping only minhashes of the tokens to keeping hashes of
all the tokens. See the vignette on minhash/LSH.
}

