% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TextReuseCorpus.R
\name{TextReuseCorpus}
\alias{TextReuseCorpus}
\alias{is.TextReuseCorpus}
\title{TextReuseCorpus}
\usage{
TextReuseCorpus(paths, dir = NULL, meta = list(),
  progress = interactive(), tokenizer = tokenize_ngrams, ...,
  hash_func = hash_string, keep_tokens = FALSE, keep_text = TRUE)

is.TextReuseCorpus(x)
}
\arguments{
\item{paths}{A character vector of paths to files to be opened.}

\item{dir}{A The path to a directory of text files.}

\item{meta}{A list with named elements for the metadata associated with this
corpus.}

\item{progress}{Display a progress bar while loading files.}

\item{tokenizer}{A function to split the text into tokens. See
\code{\link{tokenizers}}. If value is \code{NULL}, then tokenizing and
hashing will be skipped.}

\item{...}{Arguments passed on to the \code{tokenizer}.}

\item{hash_func}{A function to hash the tokens. See
\code{\link{hash_string}}.}

\item{keep_tokens}{Should the tokens be saved in the documents that are
returned or discarded?}

\item{keep_text}{Should the text be saved in the documents that are returned
or discarded?}

\item{x}{An R object to check.}
}
\description{
This is the contructor function for a \code{TextReuseCorpus}, modeled on the
virtual S3 class \code{\link[tm]{Corpus}} from the \code{tm} package. The
object is a \code{TextReuseCorpus}, which is basically a list containing
objects of class \code{\link{TextReuseTextDocument}}. Arguments are passed
along to that constructor function. You can pass either a character vector of
paths to text files using the \code{paths =} parameter, or a directory
containing text files using the \code{dir =} parameter.
}
\examples{
dir <- system.file("extdata/legal", package = "textreuse")
corpus <- TextReuseCorpus(dir = dir, meta = list("description" = "Field Codes"))
# Subset by position or file name
corpus[[1]]
names(corpus)
corpus[["ca1851-match"]]
}

