% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TextReuseTextDocument.R
\name{TextReuseTextDocument}
\alias{TextReuseTextDocument}
\title{Create an TextReuseTextDocument object}
\usage{
TextReuseTextDocument(file, n = 5, ...)
}
\arguments{
\item{file}{The path to a file, if \code{string} is not provided.}

\item{n}{The value of \code{n} in the n-grams that are created by
\link{ngrams}.}

\item{...}{Arguments are passed on to \link{readLines} if \code{file} is
  provided.}
}
\value{
An object of class \code{TextReuseTextDocument}. This object inherits
  from the virtual S3 class \code{\link[NLP]{TextDocument}} in the NLP
  package. It contains the following elements: \describe{ \item{content}{The
  text of the document.} \item{ngrams}{The document in shingled n-grams. See
  \code{\link{ngrams}}.} \item{metadata}{The document metadata, including the
  filename (if any) in \code{file}.} }
}
\description{
This function reads in a file and creates an object of class
\code{TextReuseTextDocument}. This class is used for comparing documents.
}

