% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_index.R
\name{token_index_candidates}
\alias{token_index_candidates}
\title{Extract candidate document pairs from a token index}
\usage{
token_index_candidates(index)
}
\arguments{
\item{index}{A \code{textreuse_token_index} object returned by
\code{\link{token_index}}.}
}
\value{
A \code{textreuse_candidates} data frame.
}
\description{
Extract candidate document pairs from a token index
}
